Tried to get php 5.5 working.
Alas, it was not to be. After spending at least an hour and maybe
two I made no progress in getting any output out of this cretinous
software.

It actually produces no output either in cli or fpm; cgi binary
prints a content-type header, powered by php line (thanks for this,
just what I need) and exits with 255.

The php package in squeeze has no fpm support. Luckily it is a
php 5.3.3 which phpbb supports, and I already have apache preinstalled
despite forgetting about it. Going to try using something that works.